Defeated....
After changing the entire game state to use a 1D Uint8Array, performance seems to have actually gotten worse, I'm not sure what to attribute this to, but the line clearing algorithm uses `copyWithin`, and after a few lines are cleared the performance seems to drop off a cliff.  I'm a little baffled.

This could be related to how rendering works in Preact, since I had to switch to nested `for(...)` loops instead of using map, but that shouldn't change much of anything.

I tried also moving the tailwind styles into distinct classes for the tetronimo block styles, to simply shorten the strings, but that didn't seem to change anything.

The real issue here, and I have been meaning to address it, is moving the activePiece rendering to it's own component.  This will prevent the fixed blocks on the board from repainting until a piece is set, and would also allow much smoother animation.  That shouldn't be too hard because the game model already tracks the position, it's just a matter of decoupling the UI.