Add line clear effect (#32)
Better Implementation of line-clear animation using framer-motion. This technique is applied in the same manner as the "hard drop" effect-- i.e. it's an overlay and not manipulating the board dom, except for setting the cleared row div(s) to 'opacity: 0`.

Also added sound effect for the "crash" when the remaining rows above are shifted down. (Note: this always plays, I do not yet check if every row above is clear, but that doesn't that often)

* added line clear animation and drop sound effect when animation completes

* Update output.css