Fixed infinite loop #49
I finally found the issue.  There are supposed to be two index values.  One is the index of the item in the bag (was `index`, but not is `bagIndex`).

The other is the value of the item drawn from the bag @ bagIndex (which corresponds to the index of the tetronimo).  This was given the name "drawnValue" to eliminate ambiguity in the algroithm.

This was driving me mad.